requirements.txt
asgiref==3.7.2
click==8.1.7
Django==5.0.1
django-extensions==3.2.3
djangorestframework==3.14.0
joblib==1.3.2
nltk==3.8.1
numpy==1.26.4
pandas==2.2.0
python-dateutil==2.8.2
pytz==2024.1
regex==2023.12.25
scikit-learn==1.4.0
scipy==1.12.0
six==1.16.0
sqlparse==0.4.4
threadpoolctl==3.2.0
tqdm==4.66.1
typing_extensions==4.9.0
tzdata==2023.4
